+++ src/HalfEdge2D/Controller/Delaunay/ControllerDelaunay.cpp
@@ -56,8 +56,8 @@ bool ControllerDelaunay::handleMouseMoveEvent(QMouseEvent* const event)
 
     Vec2i pos_global(event->pos().x(), event->pos().y());
     Vec2i pos_in_vp = keepInViewPort(pos_global);
-    Vec2f pos = Vec2f((float)pos_in_vp[0], (float)pos_in_vp[1]) + m_CurrentHitDistance;
-    Vec2f new_pos = invTrans(pos);
+    Vec2d pos = Vec2d((float)pos_in_vp[0], (float)pos_in_vp[1]) + m_CurrentHitDistance;
+    Vec2d new_pos = invTrans(pos);
 
     m_CurrentPoint->setPosition(new_pos);
 
@@ -94,7 +94,7 @@ bool ControllerDelaunay::handleMousePressEvent(QMouseEvent* const event)
 
     Vec2i pos_global(event->pos().x(), event->pos().y());
     Vec2i p = keepInViewPort(pos_global);
-    Vec2f p_f(p[0], p[1]);
+    Vec2d p_f(p[0], p[1]);
 
     m_CurrentPoint = getPointAtPos(invTrans(p_f));
 
@@ -210,7 +210,7 @@ bool ControllerDelaunay::handleWheelEvent(QWheelEvent* const event)
     return false;
 }
 
-Point* const ControllerDelaunay::getPointAtPos(const Vec2f& pos) const
+Point* const ControllerDelaunay::getPointAtPos(const Vec2d& pos) const
 {
     if(m_Points.empty())
         return nullptr;
@@ -225,23 +225,23 @@ Point* const ControllerDelaunay::getPointAtPos(const Vec2f& pos) const
 
 bool ControllerDelaunay::isPointInCircle(const std::array<Point* const, 3>& points, Point* const point)
 {
-    const Vec2f& p0 = m_Points[0]->getPosition();
-    const Vec2f& p1 = m_Points[1]->getPosition();
-    const Vec2f& p2 = m_Points[2]->getPosition();
-    const Vec2f& p3 = point->getPosition();
+    const Vec2d& p0 = m_Points[0]->getPosition();
+    const Vec2d& p1 = m_Points[1]->getPosition();
+    const Vec2d& p2 = m_Points[2]->getPosition();
+    const Vec2d& p3 = point->getPosition();
 
-    Vec2f p02(p0[0] * p0[0], p0[1] * p0[1]);
-    Vec2f p12(p1[0] * p1[0], p1[1] * p1[1]);
-    Vec2f p22(p2[0] * p2[0], p2[1] * p2[1]);
-    Vec2f p32(p3[0] * p3[0], p3[1] * p3[1]);
+    Vec2d p02(p0[0] * p0[0], p0[1] * p0[1]);
+    Vec2d p12(p1[0] * p1[0], p1[1] * p1[1]);
+    Vec2d p22(p2[0] * p2[0], p2[1] * p2[1]);
+    Vec2d p32(p3[0] * p3[0], p3[1] * p3[1]);
 
-    Mat3f m;
+    Mat3d m;
     m <<
         p0[0] - p3[0], p0[1] - p3[1], p02[0] - p32[0] + p02[1] - p32[1],
         p1[0] - p3[0], p1[1] - p3[1], p12[0] - p32[0] + p12[1] - p32[1],
         p2[0] - p3[0], p2[1] - p3[1], p22[0] - p32[0] + p22[1] - p32[1];
 
-    return m.determinant() > 0.0f;
+    return m.determinant() > 0.0;
 }
 
 void ControllerDelaunay::updateCircumCircles()
@@ -281,18 +281,18 @@ void ControllerDelaunay::updateCircumCircle(Circle* const circle, const std::arr
     l12.setPoint(0, points[1]->getPosition());
     l12.setPoint(1, points[2]->getPosition());
 
-    float length = 1000.0f * (l01.getLength() + l12.getLength());
+    double length = 1000.0 * (l01.getLength() + l12.getLength());
 
-    Vec2f bn0(l01.getNormal()[1], -l01.getNormal()[0]);
-    Vec2f bn1(l12.getNormal()[1], -l12.getNormal()[0]);
+    Vec2d bn0(l01.getNormal()[1], -l01.getNormal()[0]);
+    Vec2d bn1(l12.getNormal()[1], -l12.getNormal()[0]);
 
-    Vec2f sp0 = l01.pointAt(0.5f);
-    Vec2f sp1 = l12.pointAt(0.5f);
+    Vec2d sp0 = l01.pointAt(0.5);
+    Vec2d sp1 = l12.pointAt(0.5);
 
-    Vec2f bs0s = sp0 - bn0 * length;
-    Vec2f bs0e = sp0 + bn0 * length;
-    Vec2f bs1s = sp1 - bn1 * length;
-    Vec2f bs1e = sp1 + bn1 * length;
+    Vec2d bs0s = sp0 - bn0 * length;
+    Vec2d bs0e = sp0 + bn0 * length;
+    Vec2d bs1s = sp1 - bn1 * length;
+    Vec2d bs1e = sp1 + bn1 * length;
 
     Line bisect0;
     Line bisect1;
